% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{ch_to_m}
\alias{ch_to_m}
\title{Converts the input value from chains to meters}
\usage{
ch_to_m(x)
}
\arguments{
\item{x}{The value of the length of the chains unit system}
}
\value{
The value of the length of the meters unit system
}
\description{
Converts the input value from chains to meters.
}
\examples{
ch_to_m(1)

}
