% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{CT_to_FT}
\alias{CT_to_FT}
\title{Converts temperature degrees from Celsius to Fahrenheit}
\usage{
CT_to_FT(x)
}
\arguments{
\item{x}{The value of the tempreture of the Celsius tempreture unit system}
}
\value{
The value of the tempreture of the Fahrenheit unit system
}
\description{
Converts the input temperature degrees from Celsius to Fahrenheit.
}
\examples{
CT_to_FT(1)

}
