% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{FT_to_CT}
\alias{FT_to_CT}
\title{Converts temperature degrees from Fahrenheit to Celsius}
\usage{
FT_to_CT(x)
}
\arguments{
\item{x}{The value of the tempreture of the Fahreneit tempreture unit system}
}
\value{
The value of the tempreture of the Celsius unit system
}
\description{
Converts the input temperature degrees from Fahrenheit to Celsius.
}
\examples{
FT_to_CT(1)

}
