% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{ft_to_m}
\alias{ft_to_m}
\title{Converts the input value from feet to meters}
\usage{
ft_to_m(x)
}
\arguments{
\item{x}{The value of the length of the feet unit system}
}
\value{
The value of the length of the meters unit system
}
\description{
Converts the input value from feet to meters.
}
\examples{
ft_to_m(1)

}
