% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{gr_to_kg}
\alias{gr_to_kg}
\title{Converts the input value from grains to kirograms}
\usage{
gr_to_kg(x)
}
\arguments{
\item{x}{The value of the weight of the grains unit system}
}
\value{
The value of the weight of the krograms unit system
}
\description{
Converts the input value from grains to kirograms.
}
\examples{
gr_to_kg(1)

}
