% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{in_to_m}
\alias{in_to_m}
\title{Converts the input value from inches to meters}
\usage{
in_to_m(x)
}
\arguments{
\item{x}{The value of the length of the inches unit system}
}
\value{
The value of the length of the meters unit system
}
\description{
Converts the input value from inches to meters.
}
\examples{
in_to_m(1)

}
