% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{kg_to_gr}
\alias{kg_to_gr}
\title{Converts the input value from kirograms to grains}
\usage{
kg_to_gr(x)
}
\arguments{
\item{x}{The value of the weight of the kirograms unit system}
}
\value{
The value of the weight of the grains unit system
}
\description{
Converts the input value from kirograms to grains.
}
\examples{
kg_to_gr(1)

}
