% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{kg_to_lb}
\alias{kg_to_lb}
\title{Converts the input value from kirograms to pounds}
\usage{
kg_to_lb(x)
}
\arguments{
\item{x}{The value of the weight of the kirograms unit system}
}
\value{
The value of the weight of the pounds unit system
}
\description{
Converts the input value from kirograms to pounds.
}
\examples{
kg_to_lb(1)

}
