% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{kg_to_oz}
\alias{kg_to_oz}
\title{Converts the input value from kirograms to ounces}
\usage{
kg_to_oz(x)
}
\arguments{
\item{x}{The value of the weight of the kirograms unit system}
}
\value{
The value of the weight of the ounces unit system
}
\description{
Converts the input value from kirograms to ounces.
}
\examples{
kg_to_oz(1)

}
