% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{lb_to_kg}
\alias{lb_to_kg}
\title{Converts the input value from pounds to kirograms}
\usage{
lb_to_kg(x)
}
\arguments{
\item{x}{The value of the weight of the pounds unit system}
}
\value{
The value of the weight of the krograms unit system
}
\description{
Converts the input value from pounds to kirograms.
}
\examples{
lb_to_kg(1)

}
