% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{m_to_ch}
\alias{m_to_ch}
\title{Converts the input value from meters to chains}
\usage{
m_to_ch(x)
}
\arguments{
\item{x}{The value of the length of the meters unit system}
}
\value{
The value of the length of the chains unit system
}
\description{
Converts the input value from meters to chains.
}
\examples{
m_to_ch(1)

}
