% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{m_to_ft}
\alias{m_to_ft}
\title{Converts the input value from meters to feat}
\usage{
m_to_ft(x)
}
\arguments{
\item{x}{The value of the length of the meters unit system}
}
\value{
The value of the length of the feat unit system
}
\description{
Converts the input value from meters to feat.
}
\examples{
m_to_ft(1)

}
