% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{m_to_in}
\alias{m_to_in}
\title{Converts the input value from meters to inches}
\usage{
m_to_in(x)
}
\arguments{
\item{x}{The value of the length of the meters unit system}
}
\value{
The value of the length of the inches unit system
}
\description{
Converts the input value from meters to inches.
}
\examples{
m_to_in(1)

}
