% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{m_to_mi}
\alias{m_to_mi}
\title{Converts the input value from meters to miles}
\usage{
m_to_mi(x)
}
\arguments{
\item{x}{The value of the length of the meters unit system}
}
\value{
The value of the length of the miles unit system
}
\description{
Converts the input value from meters to miles.
}
\examples{
m_to_ch(1)

}
