% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{m_to_yd}
\alias{m_to_yd}
\title{Converts the input value from meters to yards}
\usage{
m_to_yd(x)
}
\arguments{
\item{x}{The value of the length of the meters unit system}
}
\value{
The value of the length of the yards unit system
}
\description{
Converts the input value from yards to meters.
}
\examples{
m_to_yd(1)

}
