% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{mi_to_m}
\alias{mi_to_m}
\title{Converts the input value from miles to meters}
\usage{
mi_to_m(x)
}
\arguments{
\item{x}{The value of the length of the miles unit system}
}
\value{
The value of the length of the meters unit system
}
\description{
Converts the input value from miles to meters.
}
\examples{
mi_to_m(1)

}
