% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{oz_to_kg}
\alias{oz_to_kg}
\title{Converts the input value from ounces to kirograms}
\usage{
oz_to_kg(x)
}
\arguments{
\item{x}{The value of the weight of the ounces unit system}
}
\value{
The value of the weight of the krograms unit system
}
\description{
Converts the input value from ounces to kirograms.
}
\examples{
oz_to_kg(1)

}
