% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{yd_to_m}
\alias{yd_to_m}
\title{Converts the input value from yards to meters}
\usage{
yd_to_m(x)
}
\arguments{
\item{x}{The value of the length of the yards unit system}
}
\value{
The value of the length of the meters unit system
}
\description{
Converts the input value from yards to meters.
}
\examples{
yd_to_m(1)

}
